Make InSituRegion a class so it can't be moved
InSituRegion hands out pointers to itself, so it needs to stay in the
same location in memory for its entire lifetime. Since it is currently
not possible in D to prevent a struct from being moved, InSituRegion
cannot be a struct. The only other option is to make it a class.

extern(C++) is used for compatibility with BetterC.