Check memory overlap without potential UB
It's not clear whether comparisons of unrelated pointers are undefined
behavior in D or not. There is evidence pointing in both directions:

* The D language spec says nothing about pointer comparisons. In
  general, where the D spec is silent, one can usually assume that the
  rules from C and C++ apply.

* Various parts of Phobos rely on comparison of potentially-unrelated
  pointers, including InSituRegion from std.experimental.allocator and
  doesPointTo from std.exception.

In the face of uncertainty, it's best to be conservative and assume that
these comparisons might be UB. As detailed in the linked blog post, the
potential UB can be avoided converting to the pointers to integers
before comparing them.