Make deallocate infallible
Previously, when handling deallocation failure, containers had two
possible choices:

1. Leak the memory.
2. Retry the deallocate call later.

However, when a container called deallocate in its destructor, only
choice (1) was possible, since the container (and by extension, the
Block(s) pointing to its allocations) was about to disappear.

Making deallocate infallible places the responsibility for handling
failure to reclaim memory entirely on the allocator. This is preferable
both because (1) the allocator is guaranteed to live at least as long as
every block allocated with it, and (2) the allocator is more likely to
know which blocks can be reclaimed and which cannot.

Allocators are not required to guarantee reclamation of deallocated
blocks. For example, InSituRegion will fail to reclaim (leak) memory
when blocks are not deallocated in reverse order of allocation.