Remove GCAllocator.deallocate
Container instantiations that don't free their memory manually should be
able to return references to their elements even in @safe code (like
GC-based containers currently do).

To do this, the generic container implementation must be able to tell
whether a particular instantiation will manually free its memory or not,
and the simplest way for it to do so is to check for the presence of the
deallocate method.